Cambiado N por int y val,maximo y tmp por long long

+++ TP1/Ejercicios/Ejercicio3/ej3.cpp
@@ -7,11 +7,11 @@ typedef long long ll;
 using namespace std;
 
 int main(int argc, char const *argv[]) {
-  long N;
-  int val;
+  int N;
+  ll val;
   cin >> N;
 
-  int maximum = 0, tmp = 0;
+  ll maximum = 0, tmp = 0;
   forn(i, N){
     cin >> val;
     tmp = max(val, tmp + val);
